build: split release.py into single-purpose scripts
release.py repeated the exact pattern #1462 asked to remove, just at
smaller scale: one script dispatching to 3 subcommands, each of which
only ever has a single caller. Splits it into version.py,
update_version.py, and update_patch_version.py, with the shared
regex/file-update helpers factored into version_files.py so the two
version-bump scripts don't duplicate that logic.
