+++ .changelog/5381.removed
@@ -1,1 +1,1 @@
-Removed `scripts/eachdist.py` in favor of tox and small purpose-built scripts for release version bumping (`scripts/release.py`) and distribution discovery (`scripts/repo_targets.py`). Renamed `eachdist.ini` to `repo.ini`.
+Removed `scripts/eachdist.py` in favor of tox and small purpose-built scripts: `scripts/version.py`, `scripts/update_version.py`, and `scripts/update_patch_version.py` for release version bumping, and `scripts/repo_targets.py` for distribution discovery. Renamed `eachdist.ini` to `repo.ini`.
