+++ .github/scripts/update-version-patch.sh
@@ -3,7 +3,7 @@
 sed -i "/\[stable\]/{n;s/version=.*/version=$1/}" repo.ini
 sed -i "/\[prerelease\]/{n;s/version=.*/version=$2/}" repo.ini
 
-./scripts/release.py update_patch_versions \
+./scripts/update_patch_version.py \
     --stable_version=$1 \
     --unstable_version=$2 \
     --stable_version_prev=$3 \
