+++ .github/workflows/prepare-patch-release.yml
@@ -25,8 +25,8 @@ jobs:
 
       - name: Set environment variables
         run: |
-          stable_version=$(./scripts/release.py version --mode stable)
-          unstable_version=$(./scripts/release.py version --mode prerelease)
+          stable_version=$(./scripts/version.py --mode stable)
+          unstable_version=$(./scripts/version.py --mode prerelease)
 
           if [[ $stable_version =~ ^([0-9]+\.[0-9]+)\.([0-9]+)$ ]]; then
             stable_major_minor="${BASH_REMATCH[1]}"
